% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotRunSharpe.R
\name{plotRunSharpe}
\alias{plotRunSharpe}
\title{Plot N-day Rolling Sharpe Ratio}
\usage{
plotRunSharpe(R, n = 252, ...)
}
\arguments{
\item{R}{return stream}

\item{n}{a lookback period}

\item{...}{additional plotting arguments}
}
\value{
a plot of the Sharpe ratio over time
}
\description{
plots an n-day running sharpe ratio
}

