% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runSharpe.R
\name{runSharpe}
\alias{runSharpe}
\title{Running Sharpe Ratio}
\usage{
runSharpe(R, n = 252, scale = NA, volFactor = 1)
}
\arguments{
\item{R}{a return series}

\item{n}{a lookback period}

\item{scale}{number of periods in a year
(daily scale = 252, monthly scale = 12, quarterly scale = 4)}

\item{volFactor}{a volatility factor -- can be raised to compute a value
biased further to volatility (volFactor > 1)
or away from volatility (volFactor < 1)
(default 1)}
}
\value{
an n-day rolling Sharpe ratio
}
\description{
Running Sharpe Ratio
}

