% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCumRets.R
\name{plotCumRets}
\alias{plotCumRets}
\title{Plot N-day Cumulative Returns}
\usage{
plotCumRets(R, n = 252, annualized = FALSE, ...)
}
\arguments{
\item{R}{a return stream}

\item{n}{a lookback amount in days (default 252)}

\item{annualized}{whether or not to annualize returns}

\item{...}{additional arguments}
}
\value{
a rolling n-day cumulative return plot in percent
}
\description{
plots a running n-day cumulative return
}

